changed screenpos struct to use floats

+++ defs.h
@@ -29,8 +29,8 @@ struct Pos {
 };
 
 struct ScreenPos {
-	int x, y, z;
-	ScreenPos(int x, int y, float z) : x(x), y(y), z(z) {}
+	float x, y, z;
+	ScreenPos(float x, float y, float z) : x(x), y(y), z(z) {}
 };
 
 struct Vector {
